JorgX39-Implementation-of-a-page-with-a-list-of-places-after-searching-and-a-reusable-search-engine-component cz 1
A raw layout was created, without support for clicking in a given place, and the code was tidied up
